requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
twilio==8.10.0
transformers==4.44.2
torch==2.1.0
Werkzeug==3.0.1
llama-cpp-python==0.2.56